+++ clang-tools-extra/clang-tidy/hooks/ArrayBufLenCheck.h
@@ -1,0 +1,36 @@
+//===--- ArrayBufLenCheck.h - clang-tidy ------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_ARRAYBUFLENCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_ARRAYBUFLENCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+/// Hook API sto_subarray requires non-empty input buffer and takes a
+/// parameter specifying the array index, whose value is limited - the
+/// sought object cannot be found if the limit is exceeded.
+class ArrayBufLenCheck : public ClangTidyCheck {
+public:
+  ArrayBufLenCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  const int64_t MAX_INDEX = 1023; // value from rippled source
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_ARRAYBUFLENCHECK_H